lib: fix queue size calculation
When next_job pointer was a smaller value than earliest_job pointer,
the calculation of the job queue size was wrong, as it did not take
into account that the pointers wrap around the array.

+++ lib/include/mb_mgr_code.h
@@ -52,7 +52,8 @@ void ADV_JOBS(int *ptr)
 __forceinline uint32_t
 get_queue_sz(IMB_MGR *state)
 {
-        const int a = (state->next_job - state->earliest_job) / sizeof(IMB_JOB);
+        const int a = (state->next_job - state->earliest_job) /
+                        (int) sizeof(IMB_JOB);
 
         return a & (IMB_MAX_JOBS-1);
 }
